Validaciones del register completas y funcionales. + Ajustes en el admin/create

+++ controllers/registerController.js
@@ -1,5 +1,6 @@
 const fs = require('fs');
 const path = require('path');
+const { validationResult } = require('express-validator');
 
 const usersFilePath = path.join(__dirname, '../data/users.json');
 const listaUsuarios = JSON.parse(fs.readFileSync(usersFilePath, 'utf-8'));
@@ -10,17 +11,25 @@ const registerController = {
     },
 
     registrar : (req, res) => {
-        const nuevoUsuario = {
-            nombre: req.body.nombre,
-            apellido: req.body.apellido,
-            email: req.body.email,
-            contraseña: req.body.contraseña,
-            confirmaContraseña: req.body.confirmaContraseña,
+
+        let errores = validationResult(req)
+        if (req.body.contraseña == req.body.confirmaContraseña && errores.isEmpty() ) {
+            const nuevoUsuario = {
+                nombre: req.body.nombre,
+                apellido: req.body.apellido,
+                email: req.body.email,
+                contraseña: req.body.contraseña,
+                confirmaContraseña: req.body.confirmaContraseña,
+            }
+            listaUsuarios.push(nuevoUsuario);
+            fs.writeFileSync(usersFilePath, JSON.stringify(listaUsuarios, null))
+            res.redirect('/login')
         }
-        listaUsuarios.push(nuevoUsuario);
-        fs.writeFileSync(usersFilePath, JSON.stringify(listaUsuarios, null))
-        res.redirect('/')
-    }
+        else {
+            return res.render ("register", { errores : errores.array(), old : req.body })
+        } 
+         
+        }
 };
 
 module.exports = registerController;
